chore: Split utils into a package.
